+++ src/org/jets3t/service/model/S3Bucket.java
@@ -53,6 +53,9 @@ public class S3Bucket extends StorageBucket {
      */
     public static final String LOCATION_SOUTH_AMERICA_EAST = LOCATION_SOUTH_AMERICA_SAO_PAULO;
 
+    public static final String LOCATION_GOVCLOUD_US_WEST = "s3-us-gov-west-1";
+    public static final String LOCATION_GOVCLOUD_FIPS_US_WEST = "s3-fips-us-gov-west-1";
+
 
     private boolean requesterPays = false;
     private boolean isRequesterPaysKnown = false;