+++ RELEASE_NOTES.markdown
@@ -28,7 +28,8 @@ RestS3Service:
 
  * Support for multiple object deletes in a single request
  * Explicit support for new S3 locations: Oregon (us-west-2),
-   South America (sa-east-1)
+   South America (sa-east-1), GovCloud US West (s3-us-gov-west-1),
+   GovCloud US West FIPS 140-2 (s3-fips-us-gov-west-1)
  * Support for server-side encryption, with per-object setting of algorithm
    and default algorithm configuration 's3service.server-side-encryption'
  * Support for Multipart Upload Part - Copy operation, to add data from